requirements.txt
streamlit>=1.30.0
lasio>=0.31
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pdfplumber>=0.10.0
plotly>=5.18.0
openpyxl>=3.1.0
Pillow>=10.0.0
anthropic>=0.39.0
google-genai>=1.0.0
